feat: クリップ動画のセグメント指定を秒とフレームでサポート
- ClipVideos.ts:
  - IClipSegmentインターフェイスにstart_frameとend_frameを追加し、秒指定とフレーム指定の両方をサポート。

- Videos.ts:
  - exportVideoClipメソッドを修正し、セグメントがフレーム指定か秒指定かを判定して適切なペイロードを構築。
  - クリップダウンロード機能をBlob経由で実装し、PWAモードでも動作するように改善。

- PlayerController.ts:
  - クリップ再生時にCMマーカーを追加しないように修正。
  - クリップの再生終了時に最初のセグメントに戻るように変更。

- vite-env.d.ts:
  - iOS Safari用のビデオ要素フルスクリーンAPIを追加。

- VideoStreamsRouter.py:
  - セグメントの正規化とバリデーションを改善し、フレームレートに基づく変換を追加。

- schemas.py:
  - ClipSegmentクラスにstart_frameとend_frameを追加し、バリデーションを強化。

- VideoStream.py:
  - SESSION_TIMEOUTを60秒に変更。

- 新しいクリップ動画ファイルを追加。

+++ server/app/streams/VideoStream.py
@@ -64,7 +64,7 @@ class VideoStream:
 
     # 録画視聴セッションが再生されていない場合にタイムアウトするまでの時間 (秒)
     # この時間が経過すると、録画視聴セッションのインスタンスは自動的に破棄される
-    SESSION_TIMEOUT: ClassVar[float] = float(10)  # 10 秒
+    SESSION_TIMEOUT: ClassVar[float] = float(60)  # 60 秒
 
     # 一度でも読み取られた HLS セグメントの最大保持数
     MAX_READED_SEGMENTS: ClassVar[int] = 10
